requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
appnope==0.1.3
    # via ipython
apscheduler==3.9.1
    # via -r requirements.in
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
async-timeout==4.0.2
    # via redis
backcall==0.2.0
    # via ipython
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.11.1
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
brotlipy==0.7.0
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
celery[redis]==5.2.7
    # via
    #   -r requirements.in
    #   celery-redbeat
celery-redbeat==2.0.0
    # via -r requirements.in
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via brotlipy
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via odfpy
deprecated==1.2.13
    # via redis
diff-match-patch==20200713
    # via django-import-export
dj-database-url==1.0.0
    # via -r requirements.in
django==3.2.15
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.13.0
    # via -r requirements.in
django-debreach==2.1.0
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==2.0.0
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-webpack-loader==1.6.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==1.0.0
    # via stack-data
faker==14.2.0
    # via -r requirements.in
google-api-core==2.10.0
    # via google-api-python-client
google-api-python-client==2.58.0
    # via -r requirements.in
google-auth==2.11.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.56.4
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.15.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.3
    # via requests
ipython==8.4.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
jsons==1.6.3
    # via -r requirements.in
kombu==5.2.4
    # via celery
lxml==4.9.1
    # via -r requirements.in
markuppy==1.14
    # via tablib
markupsafe==2.1.1
    # via werkzeug
matplotlib-inline==0.1.6
    # via ipython
numpy==1.23.2
    # via pandas
oauth2client==4.1.3
    # via -r requirements.in
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via redis
pandas==1.4.4
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.31
    # via
    #   click-repl
    #   ipython
protobuf==4.21.5
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.9.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via ipython
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via packaging
pytelegrambotapi==4.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   celery-redbeat
    #   faker
    #   pandas
python-decouple==3.6
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   -r requirements.in
    #   apscheduler
    #   celery
    #   django
    #   djangorestframework
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via tablib
redis==4.3.4
    # via
    #   celery
    #   celery-redbeat
    #   django-redis
requests==2.28.1
    # via
    #   google-api-core
    #   pytelegrambotapi
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
sentry-sdk==1.9.7
    # via -r requirements.in
six==1.16.0
    # via
    #   apscheduler
    #   asttokens
    #   click-repl
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stack-data==0.5.0
    # via ipython
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
tenacity==8.0.1
    # via celery-redbeat
traitlets==5.3.0
    # via
    #   ipython
    #   matplotlib-inline
typish==1.9.3
    # via jsons
tzdata==2022.2
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.2.2
    # via -r requirements.in
whitenoise==6.2.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
